requirements.txt
django==4.2.1
django-environ==0.10.0
gunicorn==20.1.0
djangorestframework==3.14.0
django-filter==23.2
psycopg2-binary==2.9.6
requests==2.30.0
django-storages==1.12.3
boto3==1.24.20
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
